build: reorganize project structure
