+++ include/libtrx/strings.h
