--- src/json.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "json/bson_parse.h"
-#include "json/bson_write.h"
-#include "json/json_base.h"
-#include "json/json_parse.h"
-#include "json/json_write.h"
