+++ src/json/bson_parse.c
@@ -1,6 +1,5 @@
-#include "json/bson_parse.h"
+#include "bson.h"
 
-#include "json/json_base.h"
 #include "log.h"
 #include "memory.h"
