+++ src/json/bson_write.c
@@ -1,6 +1,5 @@
-#include "json/bson_write.h"
+#include "bson.h"
 
-#include "json/json_base.h"
 #include "log.h"
 #include "memory.h"
