+++ src/json/json_base.c
@@ -1,4 +1,4 @@
-#include "json/json_base.h"
+#include "json.h"
 
 #include "memory.h"
