+++ src/json/json_parse.c
@@ -1,6 +1,5 @@
-#include "json/json_parse.h"
+#include "json.h"
 
-#include "json/json_base.h"
 #include "memory.h"
 
 struct json_parse_state_s {
